Plugin: add size throttle filter and it's unit tests
The Size Throttle Filter plugin sets the average Rate of messages size per
Interval, based on a leaky bucket and sliding window algorithm.
In the case of overflood, it will leak within a certain rate. There can be multiple
buckets as each bucket are made for a different value of name_field.
Name_field could be ,for exaple, the source (pod_name, contaner_name).
Thus, for pod kube-apiserver and alertmanager there will be two different
buckets. We can monitor only one field of the log by setting log_field,
for example, to "log". Thus we will take into account only the size of
the log field. Because after a time of no encountered new data an unused bucket
will be deleted to avoid frequent deletion we can set window_time_duration.
When we have nested fields we separate the keys by '|'.
example:
[FILTER]
	Name     		size_throttle
	Match	 		*
	Rate     		1M
	Window   		30
	Interval 		3s
	Log_field	 	log
	Name_field		kubernetes|pod_name
	Window_time_duration 	360s

input:
    [123456, {"log":"20kmessage", "kubernetes":{"pod_name":"kube-apiserver"}}]
    [123490, {"log":"30kmessage", "kubernetes":{"pod_name":"alermanager"}}]

This plugin will make two buckets (one for "kube-apiserver" and one for
"alertmanager) where will take into account only the size of the "log" field.
Log_field can be omitted, thus all of the message will be calculated.
So if the kube-apiserver starts to flood the plugin will drop the messages
only of kube-apiserver pod, not alertmanager's ones.

Signed-off-by: I330716 <vladimir.vasilev@sap.com>